+++ packages/react/src/React.ml
@@ -2,6 +2,13 @@ type 'value ref = { mutable current : 'value }
 type domRef = CallbackDomRef of (Dom.element Js.nullable -> unit) | CurrentDomRef of Dom.element Js.nullable ref
 
 module Ref = struct
+  type 'value t = 'value ref
+
+  let current ref = ref.current
+  let setCurrent ref value = ref.current <- value
+end
+
+module DOMRef = struct
   type t = domRef
   type currentDomRef = Dom.element Js.nullable ref
   type callbackDomRef = Dom.element Js.nullable -> unit
@@ -12,7 +19,6 @@ end
 
 let createRef () = { current = None }
 let useRef value = { current = value }
-let forwardRef f = f ()
 
 module Event = struct
   type 'a synthetic
@@ -71,7 +77,7 @@ module Event = struct
     let persist : 'a synthetic -> unit = fun _ -> fail "Synthetic.persist"
   end
 
-  (* let toSyntheticEvent : 'a synthetic -> Synthetic.t = i -> i *)
+  let toSyntheticEvent : 'a synthetic -> Synthetic.t = fun _ -> fail "toSyntheticEvent"
 
   module Clipboard = struct
     type tag
@@ -172,8 +178,7 @@ module Event = struct
     end)
 
     let detail : t -> int = fun _ -> fail "Pointer.detail"
-
-    (* let view : t -> Dom.window *)
+    let view : t -> Dom.window = fun _ -> fail "Pointer.view"
     let screenX : t -> int = fun _ -> fail "Pointer.screenX"
     let screenY : t -> int = fun _ -> fail "Pointer.screenY"
     let clientX : t -> int = fun _ -> fail "Pointer.clientX"
@@ -190,8 +195,7 @@ module Event = struct
     let button : t -> int = fun _ -> fail "Pointer.button"
     let buttons : t -> int = fun _ -> fail "Pointer.buttons"
     let relatedTarget : t -> target_like option = fun _ -> fail "Pointer.relatedTarget"
-
-    (* let pointerId : t -> Dom.eventPointerId *)
+    let pointerId : t -> Dom.eventPointerId = fun _ -> fail "Pointer.pointerId"
     let width : t -> float = fun _ -> fail "Pointer.width"
     let height : t -> float = fun _ -> fail "Pointer.height"
     let pressure : t -> float = fun _ -> fail "Pointer.pressure"
@@ -239,7 +243,7 @@ module Event = struct
     end)
 
     let detail : t -> int = fun _ -> fail "UI.detail"
-    (* let view : t -> Dom.window *)
+    let view : t -> Dom.window = fun _ -> fail "UI.view"
   end
 
   module Wheel = struct
@@ -356,7 +360,7 @@ module JSX = struct
     | Float of (string * string * float)
     | Style of (string * string * string) list
     | DangerouslyInnerHtml of string
-    | Ref of Ref.t
+    | Ref of domRef
     | Event of string * event
 
   let bool name jsxName value = Bool (name, jsxName, value)
@@ -402,6 +406,7 @@ module Model = struct
 end
 
 type ('props, 'return) componentLike = ?key:string -> 'props -> 'return
+and 'props component = ('props, element) componentLike
 
 and element =
   | Lower_case_element of lower_case_element
@@ -440,6 +445,8 @@ and element =
 and lower_case_element = { key : string option; tag : string; attributes : JSX.prop list; children : element list }
 and client_props = (string * element Model.t) list
 and model_value = element Model.t
+
+let forwardRef render ?key:_ props = render props Js.Nullable.null
 
 exception Invalid_children of string
 
@@ -495,6 +502,18 @@ let create_element_with_key ?key tag attributes children =
 
 let createElement = create_element_with_key ?key:None
 let createElementWithKey = create_element_with_key
+let component component = component
+
+let isValidElement = function
+  | Text _ | Int _ | Float _ | List _ | Array _ | Empty -> false
+  | Lower_case_element _ | Upper_case_component _ | Async_component _ | Client_component _ | Static _ | Writer _
+  | Fragment _ | Provider _ | Consumer _ | Suspense _ ->
+      true
+
+let jsxKeyed component props ?key () = component ?key props
+let jsx (component : 'props component) props = component ?key:None props
+let jsxs = jsx
+let jsxsKeyed = jsxKeyed
 
 let clone_component_error name =
   Printf.sprintf
@@ -529,6 +548,8 @@ module Fragment = struct
 
   let make ?key:_ props = Fragment props#children
 end
+
+module StrictMode = Fragment
 
 let fragment children = Fragment.make (Fragment.makeProps ~children ())
 
@@ -581,7 +602,7 @@ let createContext (initial_value : 'a) : 'a Context.t =
 module Suspense = struct
   let or_react_null = function None -> null | Some x -> x
 
-  let makeProps ?fallback ?children () : < fallback : element option ; children : element option > Js.t =
+  let makeProps ?children ?fallback () : < fallback : element option ; children : element option > Js.t =
     object
       method fallback = fallback
       method children = children
@@ -664,6 +685,7 @@ let useState (make_initial_value : unit -> 'state) =
   (initial_value, setState)
 
 type ('input, 'output) callback = 'input -> 'output
+type ('input, 'output) callbackAsync = 'input -> 'output Js.Promise.t
 
 let useSyncExternalStore ~subscribe:_ ~getSnapshot = getSnapshot ()
 let useSyncExternalStoreWithServer ~subscribe:_ ~getSnapshot:_ ~getServerSnapshot = getServerSnapshot ()
@@ -806,6 +828,7 @@ let useMemo3 fn _ = fn ()
 let useMemo4 fn _ = fn ()
 let useMemo5 fn _ = fn ()
 let useMemo6 fn _ = fn ()
+let useMemo7 fn _ = fn ()
 let useCallback fn = fn
 let useCallback0 fn = fn
 let useCallback1 fn _ = fn
@@ -814,6 +837,7 @@ let useCallback3 fn _ = fn
 let useCallback4 fn _ = fn
 let useCallback5 fn _ = fn
 let useCallback6 fn _ = fn
+let useCallback7 fn _ = fn
 let useReducer _ s = (s, fun _ -> ())
 let useReducerWithMapState _ s mapper = (mapper s, fun _ -> ())
 let useEffect _ = ()
@@ -824,6 +848,17 @@ let useEffect3 _ _ = ()
 let useEffect4 _ _ = ()
 let useEffect5 _ _ = ()
 let useEffect6 _ _ = ()
+let useEffect7 _ _ = ()
+let useInsertionEffect _ = ()
+let useInsertionEffect0 _ = ()
+let useInsertionEffect1 _ _ = ()
+let useInsertionEffect2 _ _ = ()
+let useInsertionEffect3 _ _ = ()
+let useInsertionEffect4 _ _ = ()
+let useInsertionEffect5 _ _ = ()
+let useInsertionEffect6 _ _ = ()
+let useInsertionEffect7 _ _ = ()
+let useLayoutEffect _ = ()
 let useLayoutEffect0 _ = ()
 let useLayoutEffect1 _ _ = ()
 let useLayoutEffect2 _ _ = ()
@@ -831,6 +866,43 @@ let useLayoutEffect3 _ _ = ()
 let useLayoutEffect4 _ _ = ()
 let useLayoutEffect5 _ _ = ()
 let useLayoutEffect6 _ _ = ()
+let useLayoutEffect7 _ _ = ()
+let useImperativeHandle0 _ _ = ()
+let useImperativeHandle1 _ _ _ = ()
+let useImperativeHandle2 _ _ _ = ()
+let useImperativeHandle3 _ _ _ = ()
+let useImperativeHandle4 _ _ _ = ()
+let useImperativeHandle5 _ _ _ = ()
+let useImperativeHandle6 _ _ _ = ()
+let useImperativeHandle7 _ _ _ = ()
+
+module Uncurried = struct
+  type ('input, 'output) callback = ('input -> 'output) Js.Fn.arity1
+
+  let wrap callback = { Js.Fn.i1 = callback }
+
+  let useState make_initial_value =
+    let state, set_state = useState make_initial_value in
+    (state, wrap set_state)
+
+  let useReducer reducer state =
+    let state, dispatch = useReducer reducer state in
+    (state, wrap dispatch)
+
+  let useReducerWithMapState reducer state mapper =
+    let state, dispatch = useReducerWithMapState reducer state mapper in
+    (state, wrap dispatch)
+
+  let useCallback callback = wrap callback
+  let useCallback0 callback = wrap callback
+  let useCallback1 callback _ = wrap callback
+  let useCallback2 callback _ = wrap callback
+  let useCallback3 callback _ = wrap callback
+  let useCallback4 callback _ = wrap callback
+  let useCallback5 callback _ = wrap callback
+  let useCallback6 callback _ = wrap callback
+  let useCallback7 callback _ = wrap callback
+end
 
 module Children = struct
   let map element fn =
@@ -882,8 +954,19 @@ module Children = struct
 end
 
 let setDisplayName _ _ = ()
+let displayName _ = None
+let startTransition callback = callback ()
 let useTransition () = (false, fun (_cb : unit -> unit) -> ())
-let useDebugValue : 'value -> ?format:('value -> string) -> unit = fun[@warning "-16"] _ ?format:_ -> ()
+
+let useTransitionAsync () =
+  ( false,
+    fun callback ->
+      callback ();
+      Js.Promise.resolve () )
+
+let act callback = callback () |> Js.Promise.resolve
+let actAsync callback = callback ()
+let useDebugValue _ ?format:_ () = ()
 let useDeferredValue value = value
 
 (* `exception Suspend of 'a Lwt`
@@ -895,6 +978,8 @@ exception Suspend of any_promise
 let suspend promise = raise (Suspend (Any_promise promise))
 
 module Experimental = struct
+  type formAction = unit
+
   let usePromise promise =
     match Lwt.state promise with
     | Sleep -> suspend promise
@@ -902,5 +987,9 @@ module Experimental = struct
     | Fail e -> raise e
     | Return v -> v
 
-  let useActionState ?permalink:_ _action state = (state, (), false)
+  let promise = usePromise
+  let useContext = useContext
+  let useOptimistic state _update = (state, fun _ -> ())
+  let useTransitionAsync () = (false, fun callback -> callback ())
+  let useActionState _action state = (state, (), false)
 end